Added gallery component
