+++ src/app/app.component.html
@@ -1,5 +1,6 @@
 <app-navbar></app-navbar>
 <app-welcome></app-welcome>
+<app-gallery></app-gallery>
 <!-- <app-story></app-story>
 <app-services></app-services>
 <app-reviews></app-reviews> -->
