+++ src/app/app.module.ts
@@ -26,6 +26,7 @@ import {
 import {
   faMap
 } from '@fortawesome/free-solid-svg-icons';
+import { GalleryComponent } from './core/pages/gallery/gallery.component';
 
 export function HttpLoaderFactory(httpClient: HttpClient) {
   return new TranslateHttpLoader(httpClient);
@@ -43,7 +44,8 @@ export function HttpLoaderFactory(httpClient: HttpClient) {
     ReviewsComponent,
     NavbarMobileComponent,
     NavbarDesktopComponent,
-    FooterComponent
+    FooterComponent,
+    GalleryComponent
   ],
   imports: [
     BrowserModule,
