+++ src/app/core/pages/gallery/gallery.component.html
@@ -1,0 +1,24 @@
+<section class="gallery">
+    <div class="container">
+        <div class="slider">
+            <div id="prev-slide" class="slider-control">prev</div>
+            <div id="next-slide" class="slider-control">next</div>
+            <div class="images">
+                <img src="../../../../assets/images/carousel/1.jpg" alt="">
+                <img src="../../../../assets/images/carousel/2.jpg" alt="">
+                <img src="../../../../assets/images/carousel/3.jpg" alt="">
+                <img src="../../../../assets/images/carousel/4.jpg" alt="">
+                <img src="../../../../assets/images/carousel/5.jpg" alt="">
+                <img src="../../../../assets/images/carousel/6.jpg" alt="">
+                <img src="../../../../assets/images/carousel/7.jpg" alt="">
+            </div>
+        </div>
+        <div class="slider-scroll">
+            <div class="track">
+                <div class="thumb">
+
+                </div>
+            </div>
+        </div>
+    </div>
+</section>
